% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cf_logl_decomp.R
\name{calc_cf_logl_decomp}
\alias{calc_cf_logl_decomp}
\title{This file is a function,calc_cf_log_t2  that represents task 2.
The basic idea behind this task is to rewrite our matrix as (Neumann) power
series: (I???M)???1 = I +M+M^2 +M^3 + ..., (20)
where, for brevity,it is referred to I as a 0-degree effect, M as a 1st degree effect, M2 as a 2nd degree effect, etc. Importantly, the above decomposition
only works if the spectral radius of M, (M), is below one, i.e., (M) < 1,
which we will have to verify below.
To conduct this decomposition exercise in our framework, we rewrite the above matrices in the following block matrix form:
c(dlogPI,dlogR)=V(I-M)^-1 and if pho(M) which is the spectral radius is below one,
we can rewrite it as summation I+M+M^2....... .
we are going to find dlogC based on summation and V(I-M)^-1 and compare
the results to the output of task1 function which is calc_cf_log_m_2 function.
\code{calc_cf_log_t2} is a function that calculates the
model for one sector assumption for a given set of data and a
shock. This function gets sample data in a matrix form
( either arbitrary matrix or real world data)
, shocks, and parameters and returns changes of welfare C based on V(I-M)^-1
and summation of I+M+M2+..... It also shows all degree effects separately.
\enumerate{
\item C_hat_inv returns results based on V(I-M)^-1.
\item C_hat_sum returns result based on V(I+M+M^2+M^3+.....) .
At the end, C_hat_inv and C_hat_sum should convert to a same number with
an infinitesimal difference.
}}
\usage{
calc_cf_logl_decomp(
  J,
  R,
  pi_I,
  pi_F,
  gamma,
  T_hat,
  tau_hat_I,
  tau_hat_F,
  epsilon,
  ktol = 1e-10,
  kmax_iter = 500
)
}
\arguments{
\item{shock}{a list containing a shock with values
\describe{
\item{T_hat}{technological shock}}
\item{tau_hat}{tariff shocks that may affect both importer and exporter}}
}
\value{
after computing all necessary variables, if the spectral radius of M
is below one, The function returns,V(I-M), I+M+M2.....,the difference,
dlogc based on V(I-M^-1 and dlogC based on summation.
Otherwise it says:  "The spectral radius is above one
the decomposition cannot be done."
and individual degree effects of M M2 ......
It also shows the number of repetitions that convergence happens.
}
\description{
The question is: Where to stop the summation I+M+M^2...?
We set a while loop to make sure we have calculated enough terms and the final answers converge.
While loop criteria is to minimize sum(square(V(I-M)^-1)-(I+M2+M3+.....)(i,j))
It can be computed based on other criteria as well if we want.
}
\details{
This script calculates welfare changes as a result of shock to one sector model:                                   #
Here we are using matrix representation of the mode.
using options() function we set digits=22 to make sure about the accuracy of the model
of  convergence.
The objective of using option function is to make sure
about the convergence of summation
}
