% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLT-package.R
\docType{package}
\name{LogLT-package}
\alias{LogLT}
\alias{LogLT-package}
\title{LogLT: Evaluate Log Linearized Trade Models}
\description{
This package is used to compare general equilibrium solutions of 
             quantiative trade models to log linearized solutions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/okrebs/LogLT}
  \item Report bugs at \url{https://github.com/okrebs/LogLT/issues}
}

}
\author{
\strong{Maintainer}: Oliver Krebs \email{oliver.b.krebs@gmail.com}

Authors:
\itemize{
  \item Helia-sudo
}

}
\keyword{internal}
