% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cf_logl.R
\name{calc_cf_logl}
\alias{calc_cf_logl}
\title{Calculate/Simulate model counterfactual in matrix one sector form}
\usage{
calc_cf_logl(
  J,
  R,
  pi_I,
  pi_F,
  gamma,
  T_hat,
  tau_hat_I,
  tau_hat_F,
  epsilon,
  use_pseudo = FALSE
)
}
\arguments{
\item{shock}{a list containing a shock with values
\describe{
\item{T_hat}{technological shock}}
\item{tau_hat}{tariff shocks that may affect both importer and exporter}}
}
\value{
after computing all X,X',A,Ap,Y1,Y1p,Y2,Y2p,B,Bp,Z1,Z1p,Z2,Z2p
Z3,Z3p,Z4p,Y3, the function returns C_hat_1 and C_hat_2 based on psuedo and normalized X'
calc_cf_log_m -> function(simple_data,shocks,parameters)
three blocks of input:
\enumerate{
\item simple_data <-
R revenue or GDP vector of size J
J number of countries
pi=matrix(c(),nrow=J) , is a vector of size J
needed to be divided into intermediate and final pi, pi_I:J\emph{J for intermediate
pi_F:J}J for final
pi_I=matrix(pi\link{1:J*J},nrow=J,ncol=J)
pi_F=matrix(pi\link[=(J*J)+1:J*J+J*J]{(J\emph{J)+1:J}J+J*J},nrow=J,ncol=J)
gamma=as.vector of size J
\item parameters <-
list(
epsilon = 3,
mobility = "immobile"
\item shock <-
list(
T_hat = matrix(vector(c(), nrow = J,ncol=1)
tau_hat=matrix(c(),nrow=4J=N,ncol=1) should be divided into two types of shocks: intermediates and final
tau_hat_I=matrix(tau_hat\link{1:(J*J)},nrow=J,ncol=J)
tau_hat_F=matrix(tau_hat\link[=(J*J)+1:(J*J)+(J*J)]{(J\emph{J)+1:(J}J)+(J*J)},nrow=J,ncol=J)
}
}
\description{
\code{calc_cf_log_m_1} is a function that calculates the
model for one sector assumption for a given set of data and a
shock. This function gets sample data in matrix form
( either arbitrary matrix or real world data)
, shocks, and parameters and returns changes of welfare C in two forms:
\enumerate{
\item C_hat_1 returns results based on X' computed based on psuedo inverse
\item C_hat_2 returns result based on X' computed based on normalized Q.
At the end, C_hat_1 and C_hat_2 should convert to a same number with
with an infinitesimal difference
}
}
\details{
This script calculates welfare changes as a result of shock to one sector model:                                   #
Here we are using matrix representation of the mode.
using options() function we set digits=22 to make sure about the accuracy of the model
of  convergence.
The objective of using option function is to make sure
about the convergence of X' using psuedo inverse in comparison
with inverse of normalized one using Q matrix which is
Q=gamma*R as it is mentioned in footnote-page 10-task 1.
}
